Rung 7c — flash through the assembly pass: delete _warp_chain.py, split is a shim
Close gap (2): the warp-tier flash no longer bypasses the pipeline with a bespoke
assembler. It now rides the standard TileGraphOp → assembly pass like every other
kernel.

- `assembly/_flash.py` (relocated from the deleted `_warp_chain.py`): `realize_flash(
  FlashSpec)` authors the flash carrier phases + `(B,H,S,D)` / GQA / mask / handoff
  geometry and materializes through the generic `assemble_carry` (the same tower path
  matmul / reduce use). `FlashSpec` carries the cells' atomized Mmas + the carrier +
  geometry.
- `TileGraphOp.flash` — a flash op the assembler realizes directly.
- `assembly/010_assemble` dispatches: `if op.flash: return realize_flash(op.flash)`.
- `split/005_warp_chain` collapses to the eligibility/offer SHIM — it detects the
  in-scope flash, atomizes the cells, reads the carrier, and emits
  `TileGraphOp(flash=FlashSpec(...))`. No assembly logic, no Graph-wiring.
- Deleted `assembly/_warp_chain.py` (`build_warp_chain_tileop` + `assemble_warp_chain` —
  the separate assembler entry).

So flash is now assembled by the one generic assembler (`assemble_carry`) reached
through the one assembly pass (`010_assemble`), with only the flash *algorithm
authoring* (irreducibly flash-specific) living in `_flash.py`. Combined with 7a/7b the
matmul / reduce / flash unification is complete at the assembler + pass level.

Green (GPU, RTX 5090): full tests/compiler/ 1685 passed, 4 skipped (-O1); flash e2e
test_flash_tensorcore_{reference,generated} at deployable -O3 (32). make lint clean.
