+++ deplodock/compiler/pipeline/passes/lowering/tile/assembly/010_assemble.py
@@ -17,6 +17,7 @@ from deplodock.compiler.graph import Graph, Node
 from deplodock.compiler.ir.tile.ir import TileGraphOp, TileOp
 from deplodock.compiler.pipeline import Pattern, RuleSkipped
 from deplodock.compiler.pipeline.passes.lowering.tile.assembly._assemble import assemble_block, assembly_ready
+from deplodock.compiler.pipeline.passes.lowering.tile.assembly._flash import realize_flash
 
 PATTERN = [Pattern("root", TileGraphOp)]
 
@@ -38,6 +39,11 @@ def rewrite(ctx: Context, root: Node, match) -> TileOp | Graph:  # noqa: ARG001
     engine splices, the same shape the structural forks (``140_atomic_free_splitk``)
     return."""
     op: TileGraphOp = root.op
+    if op.flash is not None:
+        # Warp-tier streaming flash (the offer shim ``split/005_warp_chain`` stamped the
+        # ``FlashSpec``): realize it directly through the generic carry assembler. The
+        # enumeration forks skip a flash op (it carries no logical ``tilegraph`` to tile).
+        return realize_flash(op.flash)
     if op.tilegraph is None:
         raise RuleSkipped("TileGraphOp not yet fully tiled (still a logical seed)")
     tg = op.tilegraph
