+++ deplodock/compiler/pipeline/passes/lowering/tile/assembly/_tower.py
@@ -1,8 +1,9 @@
 """Tower-building primitives for the assembly phase: the internal :class:`Role`
 label and :func:`_wrap_tower`, which wraps a body in the nested typed tile
 flavors (``GridTile`` / ``ThreadTile`` / ``RegisterTile`` / ``WarpTile`` /
-``AtomTile`` / ``SerialTile``). Used by ``assembly/_assemble`` + ``assembly/_warp_chain``
-to materialize the binding tower from a chosen ``Schedule``.
+``AtomTile`` / ``SerialTile``). Called by ``assembly/_assemble`` (``assemble_carry`` for every
+single-block kernel — pointwise / matmul / reduce / flash — and ``_assemble_group`` /
+``_assemble_multi`` for the fused / multi-launch DAGs) to materialize the binding tower.
 """
 
 from __future__ import annotations
